REF: Button position and SafeArea
Moved the button to an outside render. Also removed the SafeArea from the main view following the same logic as on transactions.js this allows to have more space for the button on the bottom of the screen, and also to see the transactions list without a white banner on the bottom. Please don't kill me :)